Use simpler BPM tracking algorithm
In this version we do without the timestamp array "samples" to
track user touches, but instead just keep an accumulator holding
the sum of all recorded time intervals between taps and a counter
tracking the number of taps.  This makes calculating the average
BPM a matter of a simple division, instead of having to add up all
previous intervals every time.

Aside from the (certainly neglectable) gain in performance and
reduction in memory usage, the resulting code is shorter and
likely much easier to understand for newcomers to Swift.